chore: Update page.tsx to open link in new tab
This commit modifies the page.tsx file to update the link to "https://bit.ly/GuideBookSDS2023" so that it opens in a new tab. By adding the "target="_blank"" attribute to the anchor tag, we ensure that the link opens in a separate browser tab, improving the user experience and preventing users from navigating away from the current page.